regle texture

+++ inc/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: tmorikaw <tmorikaw@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 04:47:12 by tmorikaw          #+#    #+#             */
-/*   Updated: 2023/10/22 09:50:52 by tmorikaw         ###   ########.fr       */
+/*   Updated: 2023/10/22 12:18:32 by tmorikaw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -110,7 +110,7 @@ typedef struct s_cub
 	int			hit;
 	int			side;
 	double		dist_to_wall;
-	double		lineheight;
+	int			lineheight;
 	int			draw_start;
 	int			draw_end;
 	double		camerax;
